сделал новое исключение перенес методы в родителя
