+++ src/service/FileBackedTaskManager.java
@@ -56,6 +56,8 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
             }
         } catch (IOException e) {
             throw new ManagerLoadException("Ошибка загрузки задачи из файла", e);
+        } catch (TaskIntersectionException e) {
+            throw new ManagerLoadException("Ошибка пересечения задач при загрузке", e);
         }
         return manager;
     }
@@ -91,9 +93,10 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
 
 
     @Override
-    public Task addTask(Task task) {
+    public Task addTask(Task task) throws TaskIntersectionException {
+
         save();
-        return super.addTask(task);
+        return  super.addTask(task);
 
     }
 
@@ -104,7 +107,7 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
     }
 
     @Override
-    public Subtask addSubtask(Subtask subtask) {
+    public Subtask addSubtask(Subtask subtask) throws TaskIntersectionException {
         save();
         return super.addSubtask(subtask);
     }
@@ -116,13 +119,13 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
     }
 
     @Override
-    public Task updateTask(Task task) {
+    public Task updateTask(Task task) throws TaskIntersectionException {
         save();
         return super.updateTask(task);
     }
 
     @Override
-    public Subtask updateSubtask(Subtask subtask) {
+    public Subtask updateSubtask(Subtask subtask) throws TaskIntersectionException {
         save();
         return super.updateSubtask(subtask);
     }
@@ -147,36 +150,35 @@ public class FileBackedTaskManager extends InMemoryTaskManager {
         return super.getHistory();
     }
 
-    public List<Task> getPrioritizedTasks() {
-        return new ArrayList<>(prioritizedTasks);
-    }
-
     // Проверка пересечения двух задач (если у обеих заданы startTime и duration)
-    public boolean tasksIntersect(Task t1, Task t2) {
-        if (t1.getStartTime() == null || t1.getDuration() == null ||
-                t2.getStartTime() == null || t2.getDuration() == null) {
-            return false;
-        }
-        LocalDateTime start1 = t1.getStartTime();
-        LocalDateTime end1 = t1.getEndTime();
-        LocalDateTime start2 = t2.getStartTime();
-        LocalDateTime end2 = t2.getEndTime();
-        return start1.isBefore(end2) && start2.isBefore(end1);
-    }
+    //перенесено в ин мемори таск менеджер
+
+//    public boolean tasksIntersect(Task t1, Task t2) {
+//        if (t1.getStartTime() == null || t1.getDuration() == null ||
+//                t2.getStartTime() == null || t2.getDuration() == null) {
+//            return false;
+//        }
+//        LocalDateTime start1 = t1.getStartTime();
+//        LocalDateTime end1 = t1.getEndTime();
+//        LocalDateTime start2 = t2.getStartTime();
+//        LocalDateTime end2 = t2.getEndTime();
+//        return start1.isBefore(end2) && start2.isBefore(end1);
+//    }
 
 
     // При добавлении или обновлении проверяем, пересекается ли задача
-    public void checkIntersection(Task task) {
-        if (task.getStartTime() == null || task.getDuration() == null) {
-            return;
-        }
-        boolean intersect = getPrioritizedTasks().stream()
-                .filter(t -> t.getId() != task.getId())
-                .anyMatch(t -> tasksIntersect(task, t));
-        if (intersect) {
-            throw new RuntimeException("Task time intersects with another task: " + task);
-        }
-    }
+    // перенесено в инмемори таск менеджер
+//    public void checkIntersection(Task task) {
+//        if (task.getStartTime() == null || task.getDuration() == null) {
+//            return;
+//        }
+//        boolean intersect = getPrioritizedTasks().stream()
+//                .filter(t -> t.getId() != task.getId())
+//                .anyMatch(t -> tasksIntersect(task, t));
+//        if (intersect) {
+//            throw new RuntimeException("Task time intersects with another task: " + task);
+//        }
+//    }
 
 
     // Пересчёт статуса и временных полей эпика на основе его подзадач
