+++ src/service/TaskIntersectionException.java
@@ -1,0 +1,7 @@
+package service;
+
+public class TaskIntersectionException extends Exception {
+    public TaskIntersectionException(String message) {
+        super(message);
+    }
+}
