+++ src/test/service/FileBackedTaskManagerTest.java
@@ -4,6 +4,7 @@ package test.service;
 import model.Task;
 import org.junit.jupiter.api.Test;
 import service.FileBackedTaskManager;
+import service.TaskIntersectionException;
 
 import java.io.File;
 import java.io.IOException;
@@ -12,7 +13,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class FileBackedTaskManagerTest {
     @Test
-    void shouldSaveTaskToFile() {
+    void shouldSaveTaskToFile() throws TaskIntersectionException {
         File file = new File("test-save.csv");
         FileBackedTaskManager manager = new FileBackedTaskManager(file);
         Task task = new Task("Test", "Description");
